Faster and more stable version with better button.
Uses a ring buffer for the snake body which saves on a lot of copies as the snake moves with the buffer instead of constantly being copied.  Drawing code is also updated to be more efficient.  Maximum snake size of 72 seems to be stable.  Also button press is detected on button down instead of button up which quite a few people complained about.